Using PreferenceFragment for settings, remove it from other activities than Main
Got stuck with handling the proper nav (up/back)...
