+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/Theme.AppCompat" >
+
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -21,6 +22,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
         <activity
             android:name=".TracksActivity"
             android:label="@string/title_activity_tracks"
@@ -30,9 +32,11 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="net.cchevalier.adnd.spotifystreamer.MainActivity" />
         </activity>
+
         <activity
             android:name=".PlayerActivity"
             android:label="@string/title_activity_player"
+            android:launchMode="singleTop"
             android:parentActivityName=".TracksActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
@@ -42,16 +46,19 @@
         <activity
             android:name=".SettingsActivity"
             android:label="@string/title_activity_settings"
+            android:launchMode="singleTop">
             android:parentActivityName=".MainActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="net.cchevalier.adnd.spotifystreamer.MainActivity" />
-        </activity>
+            />
+       </activity>
 
         <service
             android:name=".PlayerService"
             android:exported="false" >
         </service>
+
 
     </application>
