+++ app/src/main/res/layout/activity_settings.xml
@@ -1,0 +1,11 @@
+<fragment
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+
+    android:id="@+id/container_preferences_fragment"
+    android:name="net.cchevalier.adnd.spotifystreamer.fragments.SettingsFragment"
+
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+
+    />
