+++ app/src/main/res/menu/menu_settings.xml
@@ -1,0 +1,6 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+      xmlns:app="http://schemas.android.com/apk/res-auto"
+      xmlns:tools="http://schemas.android.com/tools"
+      tools:context="net.cchevalier.adnd.spotifystreamer.SettingsActivity">
+
+</menu>
